# specs/001-v1-commercial-flows/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Laumiley Modas API
  version: 1.0.0
  description: |
    Contrato da V1 comercial. Recursos de negócio usam HAL. Respostas de autenticação,
    redirecionamentos e erros usam JSON/HTTP convencional. A venda permanece no WhatsApp.
servers:
  - url: http://localhost:8080/api/v1
    description: Desenvolvimento
...
paths:
...
  /customers:
    post:
      tags: [Customers]
      summary: Cadastra um cliente com e-mail e senha
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerRegistrationRequest' }
      responses:
        '201':
          description: Cliente criado sempre com papel CLIENT
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/hal+json:
              schema: { $ref: '#/components/schemas/CustomerRepresentation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /customers/me:
    get:
      tags: [Customers]
      summary: Consulta o perfil do cliente autenticado
      operationId: getCurrentCustomer
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Perfil próprio do cliente, com telefone opcional para contas Google
          content:
            application/hal+json:
              schema: { $ref: '#/components/schemas/CustomerRepresentation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /customers/me/whatsapp:
    put:
      tags: [Customers]
      summary: Informa ou substitui o WhatsApp do cliente autenticado
      operationId: updateCurrentCustomerWhatsApp
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WhatsAppUpdateRequest' }
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/hal+json:
              schema: { $ref: '#/components/schemas/CustomerRepresentation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Autentica cliente por e-mail e senha
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Access token no corpo e refresh token opaco em cookie HttpOnly
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotaciona o refresh token e emite novo access token
      operationId: refresh
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
      security:
        - refreshCookie: []
      responses:
        '200':
          description: Tokens rotacionados
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoga o refresh token e expira seu cookie
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/XsrfToken'
      security:
        - refreshCookie: []
      responses:
        '204':
          description: Logout concluído
          headers:
            Set-Cookie: { schema: { type: string } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /products/{productId}:
    get:
      tags: [Catalog]
      summary: Consulta detalhes de produto ativo
      operationId: getPublicProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Produto com todas as imagens ordenadas
          content:
            application/hal+json:
              schema: { $ref: '#/components/schemas/ProductRepresentation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{productId}/interests:
    post:
      tags: [Interests]
      summary: Confirma interesse e cria o acompanhamento automático
      operationId: createInterest
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/IdempotencyKey'
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: Interest, Reminder e ContactRecord PENDING criados exatamente uma vez
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/hal+json:
              schema: { $ref: '#/components/schemas/InterestRepresentation' }
        '200':
          description: Repetição idempotente devolve o resultado já existente
          content:
            application/hal+json:
              schema: { $ref: '#/components/schemas/InterestRepresentation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/customers/{customerId}:
    get:
      tags: [CRM]
      summary: Consulta o contexto comercial do cliente
      operationId: getCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Cliente e elegibilidade atual para contato proativo
          content:
            application/hal+json:
              schema: { $ref: '#/components/schemas/CustomerRepresentation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: LAUMILEY_REFRESH
  headers:
    Location:
      description: URI do recurso criado
      schema: { type: string, format: uri-reference }
  parameters:
...
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CustomerId:
      name: customerId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
    XsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
  responses:
...
    BadRequest:
      description: Entrada inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    Unauthorized:
      description: Autenticação ausente ou inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    Forbidden:
      description: Operação não permitida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    NotFound:
      description: Recurso inexistente ou inacessível
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    Conflict:
      description: Conflito de unicidade, estado ou idempotência
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
...
    TooManyRequests:
      description: Limite local de tentativas excedido
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
  schemas:
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    Role:
      type: string
      enum: [CLIENT, ADMIN]
    Link:
      type: object
      required: [href]
      properties:
        href: { type: string }
        templated: { type: boolean }
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    CustomerRegistrationRequest:
      type: object
      additionalProperties: false
      required: [firstName, lastName, email, password, whatsappPhone]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 120 }
        lastName: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 8, maxLength: 200, writeOnly: true }
        whatsappPhone: { type: string, pattern: '^\+[1-9][0-9]{7,14}$' }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, writeOnly: true }
    TokenResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, role]
      properties:
        accessToken: { type: string, writeOnly: true }
        tokenType: { type: string, const: Bearer }
        expiresIn: { type: integer, description: Segundos }
        role: { $ref: '#/components/schemas/Role' }
...
    CustomerRepresentation:
      type: object
      required: [id, firstName, lastName, email, whatsappPhone, proactiveContactAuthorized, _links]
      properties:
        id: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        whatsappPhone: { type: [string, 'null'] }
        proactiveContactAuthorized: { type: boolean }
        proactiveContactEligible: { type: boolean }
        _links: { $ref: '#/components/schemas/Links' }
    WhatsAppUpdateRequest:
      type: object
      additionalProperties: false
      required: [whatsappPhone]
      properties:
        whatsappPhone: { type: string, pattern: '^\+[1-9][0-9]{7,14}$' }
...
    CategorySummary:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
...
    ProductImageRepresentation:
      type: object
      required: [id, url, primary, displayOrder, _links]
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        primary: { type: boolean }
        displayOrder: { type: integer }
        _links: { $ref: '#/components/schemas/Links' }
    ProductRepresentation:
      type: object
      required: [id, name, description, category, status, images, _links]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        category: { $ref: '#/components/schemas/CategorySummary' }
        status: { $ref: '#/components/schemas/ProductStatus' }
        images:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ProductImageRepresentation' }
        _links: { $ref: '#/components/schemas/Links' }
...
    InterestRepresentation:
      type: object
      required: [id, customerId, product, createdAt, whatsappUrl, _links]
      properties:
        id: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        product: { $ref: '#/components/schemas/ProductRepresentation' }
        createdAt: { type: string, format: date-time }
        whatsappUrl: { type: string, format: uri }
        _links: { $ref: '#/components/schemas/Links' }
...
    ApiErrorResponse:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z]+_[0-9]{3}$' }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }